+++ desktop/cmp/dash/container/DashContainerModel.ts
@@ -47,8 +47,8 @@ export interface DashContainerConfig extends DashConfig<DashContainerViewSpec, D
     /** True to include a button in each stack header showing the dash context menu. */
     showMenuButton?: boolean;
 
-    /** Size of the space between widgets */
-    borderWidth?: number;
+    /** Between items in pixels. */
+    margin?: number;
 
     /**
      * Custom settings to be passed to the GoldenLayout instance.
@@ -129,7 +129,7 @@ export class DashContainerModel extends DashModel<
     renderMode: RenderMode;
     refreshMode: RefreshMode;
     goldenLayoutSettings: PlainObject;
-    borderWidth: number;
+    margin: number;
 
     get isEmpty(): boolean {
         return this.goldenLayout && this.viewModels.length === 0;
@@ -153,7 +153,7 @@ export class DashContainerModel extends DashModel<
         contentLocked = false,
         renameLocked = false,
         showMenuButton = false,
-        borderWidth = 6,
+        margin = 6,
         goldenLayoutSettings,
         persistWith = null,
         emptyText = 'No views have been added to the container.',
@@ -182,7 +182,7 @@ export class DashContainerModel extends DashModel<
         this.contentLocked = contentLocked;
         this.renameLocked = renameLocked;
         this.showMenuButton = showMenuButton;
-        this.borderWidth = borderWidth;
+        this.margin = margin;
         this.goldenLayoutSettings = goldenLayoutSettings;
         this.emptyText = emptyText;
         this.addViewButtonText = addViewButtonText;
@@ -577,7 +577,7 @@ export class DashContainerModel extends DashModel<
                         ...this.goldenLayoutSettings
                     },
                     dimensions: {
-                        borderWidth: this.borderWidth,
+                        borderWidth: this.margin,
                         headerHeight: 25
                     }
                 },